Fixed CategoryBudgetAdd

+++ Backend/Types/Endpoint/CategoryBudgetAdd.cs
@@ -2,5 +2,5 @@
 
 public record CategoryBudgetAdd(
     int CategoryId,
-    int MonthlyBudget
+    decimal MonthlyBudget
 );
